refactor(services): external_usage_service 를 모듈 3종으로 분할
932줄 → summaries 284 · collectors 350 · service 340 · __init__ 27.
전부 800 이내.

정의 이동만 했고 본문은 한 글자도 바꾸지 않았다 (split_audit 으로 정의
15개 · 모듈레벨 할당 4개 전부 본문 일치, ruff format 이후 재실행도 0건).

## 배정 근거

의존 방향이 한 방향이다 — summaries ← service, collectors ← service 이며
summaries 와 collectors 사이에는 의존이 없다(AST 실측).

_service_instance 는 get_external_usage_service 가 global 로 재바인딩하므로
service 에 함께 남겼다. 가르면 인스턴스 사본이 분열되고 ruff·mypy·테스트를
모두 통과한 채로 두 개가 살아 있게 된다.

## httpx 패치 7건을 collectors 경로로

계획서는 '__init__.py 에 import httpx 유지' 를 제안했으나 택하지 않았다:
좁은 재노출 원칙과 충돌하고, 배럴이 그 이름을 쓰지 않아 ruff F401 이 나며,
무엇보다 httpx 를 실제로 쓰지 않는 모듈에 이름을 두는 셈이다. AST 실측상
httpx 사용은 컬렉터 3종뿐이고 셋 다 collectors.py 로 가므로 패치 타깃이
한 곳으로 모인다.

Red-Green: collectors 경로를 분할 전 형태로 되돌리면 7 failed,
복원하면 15 passed.

## 스캔 교훈 (이번에는 계획서가 맞고 내 스캔이 틀렸다)

patch( 를 정규식 앵커로 쓰면 여러 줄 호출을 놓친다. 1차 스캔이 0건을
보고했으나 계획서의 '1종 7회' 가 옳았다. Task 2 에서 같은 함정에 걸렸는데
패턴을 고치지 않아 반복됐다 — 타깃 문자열 자체만 grep 할 것.

게이트: ruff 0 · ruff format 0 · mypy 313파일 0 ·
pytest 1 failed, 1362 passed, 2 skipped (Task 3 후와 동일, 신규 테스트 없음)
